requirements.txt
fastapi
uvicorn
celery
redis
PyMuPDF
pypdf
pydantic-settings
python-multipart
yt-dlp
ffmpeg-python
pdf2docx